# backend/tests/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strom API",
    "description": "REST and WebSocket API for the Strom pipeline engine",
    "contact": {
      "name": "Per Enstedt",
      "email": "per.enstedt@eyevinn.se"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.3.26"
  },
  "paths": {
...
    "/api/flows/{flow_id}/blocks/{block_id}/player/control": {
      "post": {
        "tags": [
          "media_player"
        ],
        "summary": "Control the media player (play, pause, stop, next, previous).",
        "operationId": "control_player",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "description": "Block ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action performed"
          },
          "400": {
            "description": "Action failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/{flow_id}/blocks/{block_id}/player/goto": {
      "post": {
        "tags": [
          "media_player"
        ],
        "summary": "Go to a specific file in the playlist.",
        "operationId": "goto_file",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "description": "Block ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GotoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Goto performed"
          },
          "400": {
            "description": "Goto failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/{flow_id}/blocks/{block_id}/player/playlist": {
      "post": {
        "tags": [
          "media_player"
        ],
        "summary": "Set the playlist for a media player block.",
        "operationId": "set_playlist",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "description": "Block ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPlaylistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Playlist set"
          },
          "404": {
            "description": "Flow or block not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/{flow_id}/blocks/{block_id}/player/seek": {
      "post": {
        "tags": [
          "media_player"
        ],
        "summary": "Seek to a position in the current file.",
        "operationId": "seek_player",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "description": "Block ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Seek performed"
          },
          "400": {
            "description": "Seek failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/{flow_id}/blocks/{block_id}/player/state": {
      "get": {
        "tags": [
          "media_player"
        ],
        "summary": "Get the current state of a media player block.",
        "operationId": "get_player_state",
        "parameters": [
          {
            "name": "flow_id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "block_id",
            "in": "path",
            "description": "Block ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerStateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/flows/{id}": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Get a specific flow by ID.",
        "operationId": "get_flow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flow found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResponse"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "flows"
        ],
        "summary": "Update an existing flow (PUT alias).",
        "description": "This is an alias for the POST update endpoint, provided for RESTful API conventions.",
        "operationId": "update_flow_put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flow"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flow updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "flows"
        ],
        "summary": "Update an existing flow.",
        "operationId": "update_flow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flow"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flow updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "flows"
        ],
        "summary": "Delete a flow.",
        "operationId": "delete_flow",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Flow ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Flow deleted"
          },
          "404": {
            "description": "Flow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockInstance": {
        "type": "object",
        "description": "Block instance in a flow",
        "required": [
          "id",
          "block_definition_id",
          "properties",
          "position"
        ],
        "properties": {
          "block_definition_id": {
            "type": "string",
            "description": "Reference to the block definition"
          },
          "computed_external_pads": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ExternalPads",
                "description": "Computed external pads for this instance based on properties\nIf None, falls back to the pads from the block definition\nThis allows blocks to have dynamic pads based on their configuration"
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "Unique ID for this instance"
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "User-assigned name for this instance"
          },
          "position": {
            "$ref": "#/components/schemas/Position",
            "description": "Position in the visual editor"
          },
          "properties": {
            "type": "object",
            "description": "Property values for this instance",
            "additionalProperties": {
              "$ref": "#/components/schemas/PropertyValue"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "runtime_data": {
            "type": [
              "object",
              "null"
            ],
            "description": "Runtime data (not persisted to storage, only available when flow is running)\nUsed for things like generated SDP for AES67 blocks",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "Element": {
        "type": "object",
        "description": "Represents a GStreamer element instance in a flow.",
        "required": [
          "id",
          "element_type",
          "position"
        ],
        "properties": {
          "element_type": {
            "type": "string",
            "description": "GStreamer element type (e.g., \"videotestsrc\", \"x264enc\", \"filesink\")"
          },
          "id": {
            "$ref": "#/components/schemas/String",
            "description": "Unique identifier for this element instance"
          },
          "pad_properties": {
            "type": "object",
            "description": "Pad properties (pad_name -> property_name -> value)",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "$ref": "#/components/schemas/PropertyValue"
              },
              "propertyNames": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "position": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "number",
                "format": "float"
              },
              {
                "type": "number",
                "format": "float"
              }
            ],
            "description": "Display position in the visual editor (x, y)"
          },
          "properties": {
            "type": "object",
            "description": "Element properties as key-value pairs",
            "additionalProperties": {
              "$ref": "#/components/schemas/PropertyValue"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response.",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "ExternalPads": {
        "type": "object",
        "description": "External pads that the block exposes",
        "required": [
          "inputs",
          "outputs"
        ],
        "properties": {
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalPad"
            },
            "description": "Input pads (mapped to internal element pads)"
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalPad"
            },
            "description": "Output pads (mapped to internal element pads)"
          }
        }
      },
      "Flow": {
        "type": "object",
        "description": "A complete GStreamer pipeline definition.\n\nA flow represents a named, configured GStreamer pipeline that can be\nstarted, stopped, and persisted.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockInstance"
            },
            "description": "Block instances in this flow"
          },
          "elements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Element"
            },
            "description": "Elements in this flow"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for this flow"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Link"
            },
            "description": "Links between element pads and/or block external pads"
          },
          "name": {
            "type": "string",
            "description": "Human-readable name"
          },
          "properties": {
            "$ref": "#/components/schemas/FlowProperties",
            "description": "Flow configuration properties"
          },
          "state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PipelineState",
                "description": "Current runtime state (persisted to storage for automatic restart)"
              }
            ]
          }
        }
      },
...
      "FlowResponse": {
        "type": "object",
        "description": "Response containing a single flow.",
        "required": [
          "flow"
        ],
        "properties": {
          "flow": {
            "$ref": "#/components/schemas/Flow"
          }
        }
      },
...
      "GotoRequest": {
        "type": "object",
        "description": "Request to go to a specific file.",
        "required": [
          "index"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "description": "File index (0-based)",
            "minimum": 0
          }
        }
      },
...
      "Link": {
        "type": "object",
        "description": "A link between two element pads.\n\nFor API/serialization compatibility, this uses strings. Internally converted to ElementPadRef.",
        "required": [
          "from",
          "to"
        ],
        "properties": {
          "from": {
            "type": "string",
            "description": "Source element and pad (format: \"element_id\" or \"element_id:pad_name\")"
          },
          "to": {
            "type": "string",
            "description": "Destination element and pad (format: \"element_id\" or \"element_id:pad_name\")"
          }
        }
      },
...
      "PipelineState": {
        "type": "string",
        "description": "GStreamer pipeline state.\n\nThese states correspond to the GStreamer GST_STATE enum.",
        "enum": [
          "Null",
          "Ready",
          "Paused",
          "Playing"
        ]
      },
      "PlayerAction": {
        "type": "string",
        "description": "Player control action.",
        "enum": [
          "play",
          "pause",
          "stop",
          "next",
          "previous"
        ]
      },
      "PlayerControlRequest": {
        "type": "object",
        "description": "Request to control the media player.",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PlayerAction"
          }
        }
      },
      "PlayerStateResponse": {
        "type": "object",
        "description": "Response with the current player state.",
        "required": [
          "state",
          "position_ns",
          "duration_ns",
          "current_file_index",
          "total_files",
          "playlist",
          "loop_playlist"
        ],
        "properties": {
          "current_file": {
            "type": [
              "string",
              "null"
            ],
            "description": "Current file path/URI"
          },
          "current_file_index": {
            "type": "integer",
            "description": "Current file index (0-based)",
            "minimum": 0
          },
          "duration_ns": {
            "type": "integer",
            "format": "int64",
            "description": "Total duration in nanoseconds",
            "minimum": 0
          },
          "loop_playlist": {
            "type": "boolean",
            "description": "Whether playlist loops"
          },
          "playlist": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Full playlist"
          },
          "position_ns": {
            "type": "integer",
            "format": "int64",
            "description": "Current position in nanoseconds",
            "minimum": 0
          },
          "state": {
            "type": "string",
            "description": "Current playback state: \"playing\", \"paused\", \"stopped\""
          },
          "total_files": {
            "type": "integer",
            "description": "Total number of files in playlist",
            "minimum": 0
          }
        }
      },
      "Position": {
        "type": "object",
        "description": "Position in the visual editor",
        "required": [
          "x",
          "y"
        ],
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "PropertyValue": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "integer",
            "format": "int64"
          },
          {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          {
            "type": "number",
            "format": "double"
          },
          {
            "type": "boolean"
          }
        ],
        "description": "Property value that can be various types.\n\nGStreamer properties can be strings, numbers, booleans, enums, etc."
      },
...
      "SeekRequest": {
        "type": "object",
        "description": "Request to seek to a position.",
        "required": [
          "position_ns"
        ],
        "properties": {
          "position_ns": {
            "type": "integer",
            "format": "int64",
            "description": "Position in nanoseconds",
            "minimum": 0
          }
        }
      },
...
      "SetPlaylistRequest": {
        "type": "object",
        "description": "Request to set the playlist.",
        "required": [
          "files"
        ],
        "properties": {
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of file URIs (e.g., \"file:///path/to/video.mp4\")"
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
    }
  },
...
}